Refactor settings UI to use settings-factory module
Moved settings UI logic from index.js to a new src/ui/settings-factory.js module. This modularizes the code for creating, updating, and styling the settings UI, improving maintainability and separation of concerns. Also added a new CSS file for settings UI styles.
